backend/requirements.txt
crewai[tools]>=0.108.0
google-genai>=1.0.0
Pillow>=10.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
supabase>=2.13.0
